fix usage of vec3 instead of dvec3

+++ CesiumGeometry/src/OrientedBoundingBox.cpp
@@ -102,7 +102,7 @@ double OrientedBoundingBox::computeDistanceSquaredToPosition(
 
 // TODO: add test for this
 bool OrientedBoundingBox::contains(const glm::dvec3& position) const noexcept {
-  glm::vec3 localPosition = this->_inverseHalfAxes * position;
+  glm::dvec3 localPosition = this->_inverseHalfAxes * position;
   return glm::abs(localPosition.x) <= 1.0 && glm::abs(localPosition.y) <= 1.0 &&
          glm::abs(localPosition.z) <= 1.0;
 }
